+++ app/src/main/java/com/example/answer/github/view/SearchFragment.kt
@@ -1,25 +1,26 @@
-package com.example.answer.github.fragments
+package com.example.answer.github.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.annotation.NonNull
 import androidx.annotation.Nullable
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.answer.R
-import com.example.answer.databinding.FragmentLikeBinding
-import com.example.answer.github.view.GithubActivity
+import com.example.answer.databinding.FragmentSearchBinding
 import com.example.answer.github.viewmodel.GithubViewModel
-import com.example.answer.github.view.GithubLikeAdapter
+import com.example.answer.github.view.adapter.GithubSearchAdapter
 
-class LikeFragment : Fragment() {
+
+class SearchFragment : Fragment() {
     private var githubViewModel: GithubViewModel? = null
-    private lateinit var binding: FragmentLikeBinding
+    private lateinit var binding: FragmentSearchBinding
     private lateinit var view: GithubActivity
-    private lateinit var adapter: GithubLikeAdapter
+    private lateinit var adapter: GithubSearchAdapter
 
     override fun onCreate(@Nullable savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -32,9 +33,10 @@ class LikeFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
+
         val roomDetailLayoutManager = LinearLayoutManager(view)
 
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_like, container, false)
+        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_search, container, false)
 
         binding.searchRecyclerView.adapter = adapter
         binding.searchRecyclerView.layoutManager = roomDetailLayoutManager
@@ -43,19 +45,32 @@ class LikeFragment : Fragment() {
         return binding.root
     }
 
+    override fun onViewCreated(@NonNull view: View, @Nullable savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.lifecycleOwner = this
+        binding.viewModel = githubViewModel
+    }
+
+    fun clearText() {
+        binding.searchEditText.text.clear()
+    }
 
     fun setContext(view: GithubActivity){
         this.view = view
     }
 
-    fun setAdapter(adapter : GithubLikeAdapter) {
+    fun setAdapter(adapter : GithubSearchAdapter) {
         this.adapter = adapter
     }
 
+    fun getString() : String {
+        return binding.searchEditText.text.toString()
+    }
 
     companion object {
-        fun newInstance(): LikeFragment {
-            return LikeFragment()
+        fun newInstance(): SearchFragment {
+            return SearchFragment()
         }
     }
 }
