+++ app/src/main/res/layout/fragment_like.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".github.fragments.LikeFragment">
+    tools:context=".github.view.LikeFragment">
 
     <data>
         <variable
